Tree Command (#32)
* shell: tree command

This command walks the file system, creates a tree graph of it's structure,
and prints that graph to the console. Still needs some work.

* shell: finish `tree` cmd

Accepts a directory argument and prints a loading message to stderr
if tree gen is taking more than a second. Tree gen will also ignore
`.git/` and `/sys/dev/` directories unless they're the root node.

* terminal: increase scrollback to 500