Download command (#49)
* shell: `dl` command can download files

My initial implementation of getting (from indexedfs) or creating a blob and
passing that didn't work out (duplex issue). I'm not sure if this approach
is any slower/faster though. Just something to come back to later.

* shell: moved `dl` to it's own file and added support for directories

Both `archive/tar` and `archive/zip` have issues storing subdirectories.
Weirdly this only happens if not using compression, otherwise it works fine.

I'm not sure if this is user error or an issue in the stdlib, but as a
work-around I've elected to always compress when downloading directories.

* Update host.js

drop console.log

---------

Co-authored-by: Jeff Lindsay <progrium@gmail.com>