indexedfs: properly communicate errors from JS helper to Go
Adds an addition property to errors reported from `indexedfs.js`, containing
the name of a Go error to join with the returned `js.Error`. This allows
for more idiomatic error checking. It also ensures indexedfs errors can
be resolved to an ERRNO value in `jsutil.go:ToJSError()`, which is required
for `os` calls to work in userspace processes.

I've also cleaned up some unecessary code and error handling.