kernel: fix issue where standalone bootloader can't load terminal.
Since the terminal needs some files of the kernel under web we embed
the kernel source and copy it into the filesystem. The terminal itself
also needs to be copied in. This also adds back the shell binary so
that first boots are fast, but its important the binary is added after
the source so it doesn't think it needs to be built based on mtime.