build: cache pkg in `/sys/build` instead of temp dir
Using the temp dir isn't as practical considering we have to reload
to update the shell binary, causing needless unpacking every time.
In the future this may become a non-issue with process reloading.

That being said, build-pkg data doesn't change very often and it makes sense
to have a semi-permanent cache.