Procfs and Kernel refactor (#116)
* kernel: remove `engine` package

I'm finding the `engine` package to be more of a hinderance than
a benefit. It's pretty confusing to the uninitiated, and I need more
flexibility and control.

This commit removes all usages, and allows us to better reorganize the kernel.

* procfs: initial read-only implementation

PID files contain the process info at the time of opening. Polling is required
to keep up to date process information, although this wont work as of now since
`fileInfo.ModTime` always returns the Unix Epoch.

Reading a PID file returns the `proc.Process` serialized to JSON. The format
can be changed if required. In the future I'd like to add more structured
data and queries using the filesystem.

* procfs: `Remove`

* fs: clean nested ifs