boot: fetch `wanix-initfs.gz` and unpack into `/sys/*` on first boot
This commit makes some changes to the structure of `boot/initfs/`
and how initfs files are copied. Instead of hardcoding which files go
where, `fs.go` just unpacks the contents of initfs into `/sys/*`.

It's basically running the equivalent of:
```sh
cd /sys/ && \
get wanix-initfs.gz && \
gzip -d wanix-initfs.gz && \
tar -xf wanix-initfs
```

This gives us direct and simple control of file placement.
And if in the future we want to put things outside of `/sys`,
we can just unpack to `/` and update the build paths. Though
it makes sense to me that initfs would only include system related files.